reorganize the grpc code a little bit more
